// models/EquipmentModels.tsx
export interface EquipmentItem {
  id: string;
  equipmentNumber: string;
  description: string;
  sortField: string;
  sortName: string;
  functionalLocationId: string;
  functionalLocationName: string;
  plannerGroupId: string;
  plannerGroupName: string;
  catalogProfileCode: string;
  objectTypeCode: string;
}